remove stub secrets from the index
Until now, secrets were committed as empty files
that were to be populated before a build using
`nix run .#nixosConfigurations.<name>.config.passthru.bootstrap-decrypt-secrets`
so builds could be pure (Nix flakes only consider files in the index).

To ensure no actual secrets were committed, a pre-commit hook checked
all bootstrap secrets defined in all NixOS configurations to see
whether it has added changes in the index and would warn you if so.

In practice however, this was annoying because all these secret files
cluttered `git status` and `git diff` output, thereby slowing down
development pace.

It is now required to build with `--impure` if secrets are needed.

Purity of builds was not very beneficial anyway as nobody is able to
replicate a build that requires secrets so caching is also compromised.

This has the added benefit that secrets can be wiped off disk
using `git clean` to keep the time they are exposed to a minimum.