colmena/nodes/node-0: use DHCP only in initrd
DHCP in stage 2 works but can take a while after a reboot.
I think this is because it acquires an IP lease in the initrd
(because `boot.zfs.unlockEncryptedPoolsViaSSH` enables networking),
and then, due to `boot.initrd.network.flushBeforeStage2`, forgets it
and tries to acquire it again, which the router's DHCP server
does not allow (even though the MAC address does not change)
because the lease acquired in the initrd is still active.
It then takes 30 minutes (as configured in the router)
for the lease to expire before the machine can acquire a new one.

Instead of trying to acquire a new lease in stage 2
we can simply keep using the lease from stage 1
by not flushing before stage 2 and disabling DHCP in stage 2.